Moved request and response objects under transfer_objs
