+++ api/portfolio_analyzer.py
@@ -1,7 +1,7 @@
 import yfinance as yf
 import pandas as pd
 import time
-from .portfolio_response import Portfolio
+from api.transfer_objs.portfolio_response import Portfolio
 import time
 import contextlib, io
